+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add pre-commit
+- Add failing probe for testing error handling
+
+### Fixed
+- Handle exceptions from probes more gracefully
 
 ## [0.2.0] - 2023-05-24
 ### Added